Changed barrel_cortex into a module
